* src/resources directory * Interface for Svn (TileService)
