--- GshhsHeader.java
@@ -1,8 +1,0 @@
-
-public class GshhsHeader {
-  int		id, n;
-  byte		level;
-  boolean	greenwich;
-  int		west, east, south, north,
-		area, areaFull, container, ancestor;
-}
