+++ resources/wdb_borders_l.b
