+++ src/GshhsHeader.java
@@ -1,0 +1,8 @@
+
+public class GshhsHeader {
+  int		id, n;
+  byte		level;
+  boolean	greenwich;
+  int		west, east, south, north,
+		area, areaFull, container, ancestor;
+}
